# Use unsafe labels when reading the index header