--- a/pkg/storegateway/indexheader/binary_reader_test.go
+++ b/pkg/storegateway/indexheader/binary_reader_test.go
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
 package indexheader
 
 import (