#### What this PR does

When we're reading the index header, we're creating new labels.Label struct by allocating new strings all the time, but in most of the cases we just want to compare whether the label name has changed or not.

For that purpose, we can yolo the underlying buffer, avoiding that allocation.

There are two cases when we need to unyolo:
- When we create the entry in the offsets map (that reference will be stored)
- When we store the label value.

Benchmark results are quite promising:

```
name                                old time/op    new time/op    delta
BinaryReader_LargerBlock/benchmark    1.95ms ± 9%    1.17ms ± 5%  -39.99%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
BinaryReader_LargerBlock/benchmark    1.06MB ± 0%    0.16MB ± 0%  -85.16%  (p=0.000 n=4+5)

name                                old allocs/op  new allocs/op  delta
BinaryReader_LargerBlock/benchmark     20.1k ± 0%      1.3k ± 0%  -93.42%  (p=0.008 n=5+5)
```

#### Which issue(s) this PR fixes or relates to

Follow up on https://github.com/grafana/mimir/pull/3393

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`