@colega 2022-11-11
Note that this is being superseded by https://github.com/grafana/mimir/pull/3436